Implement a small HTTP server instea of relying on the WS server
The WebSocket server expects all headers to be already in the read buffer
when esthablishing the channel, and this seems to be OK for most WS clients
however this isn't the case for XHR sent from Safari.

Instead of trying to fix the WS server and hack patches on top, have a
separate HTTP server for client files and pass the pid through an
explicit message on the WebSocket.